refactor: extract generateSsrEntry and generateBrowserEntry into entries/
Moves generateSsrEntry (app-ssr-entry.ts) and generateBrowserEntry
(app-browser-entry.ts) out of the 3,215-line app-dev-server.ts into a new
entries/ directory, as described in issue #253.

app-dev-server.ts re-exports both functions so all existing import sites
(including entry-templates.test.ts) continue to work without changes.

All 10 entry-template snapshot tests pass unchanged — generated output is
byte-for-byte identical to before the move.
